[expo-cli] restructure eas builds files
