+++ packages/expo-cli/src/commands/eas-build/types.ts
@@ -1,0 +1,27 @@
+import { Platform } from '@expo/build-tools';
+
+export enum BuildCommandPlatform {
+  ANDROID = 'android',
+  IOS = 'ios',
+  ALL = 'all',
+}
+
+export enum BuildStatus {
+  IN_QUEUE = 'in-queue',
+  IN_PROGRESS = 'in-progress',
+  ERRORED = 'errored',
+  FINISHED = 'finished',
+}
+
+export interface BuildInfo {
+  id: string;
+  status: BuildStatus;
+  platform: Platform;
+  createdAt: string;
+  artifacts?: BuildArtifacts;
+}
+
+interface BuildArtifacts {
+  buildUrl?: string;
+  logsUrl: string;
+}
